fix some more tests
