+++ timelock-agent/src/test/java/com/palantir/timelock/config/TimeLockPersistenceInvariantsTest.java
@@ -15,8 +15,6 @@
  */
 package com.palantir.timelock.config;
 
-import static org.assertj.core.api.Assertions.assertThatCode;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
@@ -58,60 +56,6 @@ public class TimeLockPersistenceInvariantsTest {
         verify(mockFile, atLeastOnce()).isDirectory();
     }
 
-    @Test
-    public void throwsIfConfiguredToBeNewServiceWithExistingDirectory() throws IOException {
-        File mockFile = getMockFileWith(true, true);
-
-        assertFailsToBuildConfiguration(
-                PaxosInstallConfiguration.builder().dataDirectory(mockFile).isNewService(true));
-    }
-
-    @Test
-    public void throwsIfConfiguredToBeExistingServiceWithoutDirectory() throws IOException {
-        File mockFile = getMockFileWith(false, true);
-
-        assertFailsToBuildConfiguration(
-                PaxosInstallConfiguration.builder().dataDirectory(mockFile).isNewService(false));
-    }
-
-    @Test
-    public void newServiceByClusterBootstrapConfigurationFailsIfDirectoryExists() throws IOException {
-        File mockFile = getMockFileWith(true, true);
-
-        ClusterConfiguration differentClusterConfig = ImmutableDefaultClusterConfiguration.builder()
-                .localServer(SERVER_A)
-                .addKnownNewServers(SERVER_A)
-                .cluster(PartialServiceConfiguration.of(ImmutableList.of(SERVER_A, "b", "c"), Optional.empty()))
-                .build();
-
-        assertThatThrownBy(TimeLockInstallConfiguration.builder()
-                        .cluster(differentClusterConfig)
-                        .paxos(PaxosInstallConfiguration.builder()
-                                .dataDirectory(mockFile)
-                                .isNewService(false)
-                                .build())::build)
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    public void newServiceByClusterBootstrapConfigurationSucceedsIfDirectoryDoesNotExist() throws IOException {
-        File mockFile = getMockFileWith(false, true);
-
-        ClusterConfiguration differentClusterConfig = ImmutableDefaultClusterConfiguration.builder()
-                .localServer(SERVER_A)
-                .addKnownNewServers(SERVER_A)
-                .cluster(PartialServiceConfiguration.of(ImmutableList.of(SERVER_A, "b", "c"), Optional.empty()))
-                .build();
-
-        assertThatCode(TimeLockInstallConfiguration.builder()
-                        .cluster(differentClusterConfig)
-                        .paxos(PaxosInstallConfiguration.builder()
-                                .dataDirectory(mockFile)
-                                .isNewService(false)
-                                .build())::build)
-                .doesNotThrowAnyException();
-    }
-
     private File getMockFileWith(boolean isDirectory, boolean canCreateDirectory) throws IOException {
         File mockFile = mock(File.class);
         when(mockFile.mkdirs()).thenReturn(canCreateDirectory);
@@ -128,13 +72,5 @@ public class TimeLockPersistenceInvariantsTest {
                 .cluster(CLUSTER_CONFIG)
                 .paxos(installConfiguration)
                 .build();
-    }
-
-    private void assertFailsToBuildConfiguration(ImmutablePaxosInstallConfiguration.Builder configBuilder) {
-        PaxosInstallConfiguration installConfiguration = configBuilder.build();
-        assertThatThrownBy(TimeLockInstallConfiguration.builder()
-                        .cluster(CLUSTER_CONFIG)
-                        .paxos(installConfiguration)::build)
-                .isInstanceOf(IllegalArgumentException.class);
     }
 }
